Refactor: 폴더명 변경으로 경로 수정
