+++ pages/mydashboard/index.tsx
@@ -1,4 +1,4 @@
-import NavMyDashboard from "@/components/Gnb/NavMyDashboard";
+import NavMyDashboard from "@/components/Navbar/NavMyDashboard";
 
 export default function MyDashBoard() {
   return (
